Compacted simple scalars like EventID into a macro
