+++ src/adapter/graphql/src/scalars/event_id.rs
@@ -15,44 +15,6 @@ use crate::prelude::*;
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct EventID(evs::EventID);
+simple_scalar!(EventID, evs::EventID);
 
 /////////////////////////////////////////////////////////////////////////////////////////
-
-impl From<evs::EventID> for EventID {
-    fn from(value: evs::EventID) -> Self {
-        EventID(value)
-    }
-}
-
-impl Into<evs::EventID> for EventID {
-    fn into(self) -> evs::EventID {
-        self.0
-    }
-}
-
-impl Deref for EventID {
-    type Target = evs::EventID;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-#[Scalar]
-impl ScalarType for EventID {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        if let Value::String(s) = &value {
-            match s.parse() {
-                Ok(i) => Ok(Self(evs::EventID::new(i))),
-                Err(_) => Err(InputValueError::expected_type(value)),
-            }
-        } else {
-            Err(InputValueError::expected_type(value))
-        }
-    }
-
-    fn to_value(&self) -> Value {
-        Value::String(self.0.to_string())
-    }
-}
