+++ src/adapter/graphql/src/scalars/mod.rs
@@ -40,3 +40,49 @@ pub(crate) use os_path::*;
 pub(crate) use pagination::*;
 pub(crate) use task_id::*;
 pub(crate) use task_status_outcome::*;
+
+macro_rules! simple_scalar {
+    ($name: ident, $source_type: ty) => {
+        #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+        pub struct $name($source_type);
+
+        impl From<$source_type> for $name {
+            fn from(value: $source_type) -> Self {
+                $name(value)
+            }
+        }
+
+        impl Into<$source_type> for $name {
+            fn into(self) -> $source_type {
+                self.0
+            }
+        }
+
+        impl Deref for $name {
+            type Target = $source_type;
+            fn deref(&self) -> &Self::Target {
+                &self.0
+            }
+        }
+
+        #[Scalar]
+        impl ScalarType for $name {
+            fn parse(value: Value) -> InputValueResult<Self> {
+                if let Value::String(s) = &value {
+                    match s.parse() {
+                        Ok(i) => Ok(Self(<$source_type>::new(i))),
+                        Err(_) => Err(InputValueError::expected_type(value)),
+                    }
+                } else {
+                    Err(InputValueError::expected_type(value))
+                }
+            }
+
+            fn to_value(&self) -> Value {
+                Value::String(self.0.to_string())
+            }
+        }
+    };
+}
+
+pub(crate) use simple_scalar;
